Version 5 of Detect Adjacent Properties: optimized spatial grid fully functional, with automatic sizing and one-cell assignment per property. Improved performance and adjacency detection accuracy.

+++ src/main/java/DetectAdjacentProperties/AdjacentPropertyPair.java
@@ -9,11 +9,11 @@ public class AdjacentPropertyPair {
         this.propertyId2 = propertyId2;
     }
 
-    public int getPropertyId1() {
+    public long getPropertyId1() {
         return propertyId1;
     }
 
-    public int getPropertyId2() {
+    public long getPropertyId2() {
         return propertyId2;
     }
